Split intialization tasks into separate files
