+++ src/initialization/index.js
@@ -1,0 +1,13 @@
+const populateKeyValues = require('./populateKeyValues.js')
+const populateSchedules = require('./populateSchedules')
+const setupCommands = require('./setupCommands.js')
+const setupModels = require('./setupModels.js')
+const setupRateLimiters = require('./setupRateLimiters.js')
+
+module.exports = {
+  populateKeyValues,
+  populateSchedules,
+  setupCommands,
+  setupModels,
+  setupRateLimiters
+}
