+++ src/structs/Client.js
@@ -9,7 +9,7 @@ const ArticleMessage = require('./ArticleMessage.js')
 const Feed = require('./db/Feed.js')
 const FeedData = require('./FeedData.js')
 const ArticleMessageRateLimiter = require('./ArticleMessageRateLimiter.js')
-const initialize = require('../util/initialization.js')
+const initialize = require('../initialization/index.js')
 const getConfig = require('../config.js').get
 const createLogger = require('../util/logger/create.js')
 const connectDb = require('../util/connectDatabase.js')
