+++ src/structs/ClientManager.js
@@ -7,7 +7,7 @@ const Supporter = require('./db/Supporter.js')
 const ScheduleManager = require('./ScheduleManager.js')
 const EventEmitter = require('events').EventEmitter
 const maintenance = require('../maintenance/index.js')
-const initialize = require('../util/initialization.js')
+const initialize = require('../initialization/index.js')
 const createLogger = require('../util/logger/create.js')
 const ipc = require('../util/ipc.js')
 const configuration = require('../config.js')
