+++ src/tests/models/middleware/int_FilteredFormat.test.js
@@ -1,5 +1,5 @@
 const FilteredFormatModel = require('../../../models/FilteredFormat.js')
-const initialize = require('../../../util/initialization.js')
+const initialize = require('../../../initialization/index.js')
 const mongoose = require('mongoose')
 
 const dbName = 'test_int_middleware_format'
