+++ src/tests/models/middleware/int_Subscriber.test.js
@@ -1,5 +1,5 @@
 const SubscriberModel = require('../../../models/Subscriber.js')
-const initialize = require('../../../util/initialization.js')
+const initialize = require('../../../initialization/index.js')
 const mongoose = require('mongoose')
 
 const dbName = 'test_int_middleware_subscriber'
