+++ src/tests/structs/db/int_Supporter.database.test.js
@@ -3,7 +3,7 @@ const config = require('../../../config.js')
 const mongoose = require('mongoose')
 const Supporter = require('../../../structs/db/Supporter.js')
 const Patron = require('../../../structs/db/Patron.js')
-const initialize = require('../../../util/initialization.js')
+const initialize = require('../../../initialization/index.js')
 const dbName = 'test_int_supporters'
 const CON_OPTIONS = {
   useNewUrlParser: true,
